リソース詳細整える(不完全)
