+++ app/features/resource/detail/Presenter/index.tsx
@@ -1,15 +1,11 @@
-import { Calendar, MapPin, User } from "lucide-react";
 import type React from "react";
 import { type JSX, useEffect, useState } from "react";
 import { useLocation } from "react-router";
-import type { Additional } from "~/shared/generated/fastAPI.schemas";
-import { toFormulas } from "~/shared/lib/formula";
 import { cn } from "~/shared/lib/utils";
 import { useResourceDetail } from "../Context";
-import RefLinkSentence from "../LinkedSentence";
-import Relations from "../Relations";
 import { useTraceMemory } from "../TraceMemory/hooks";
 import { getHeadingLevel, toAdjacent } from "../util";
+import DefPresenter from "./DefPresenter";
 type Props = {
   id: string;
   prefix?: React.ReactNode;
@@ -26,7 +22,6 @@ export default function Presenter({ id, prefix }: Props) {
   const [isVisible, setIsVisible] = useState(false);
   const location = useLocation();
   const isActive = location.hash === `#${adj.kn.uid}`;
-  const refs = adj.refers();
 
   useEffect(() => {
     if (level === 0) {
@@ -55,91 +50,9 @@ export default function Presenter({ id, prefix }: Props) {
           "bg-yellow-100 text-neutral-800 dark:bg-yellow-800/30 dark:text-white",
       )}
     >
-      {/* {adj.referreds().map((ref) => ref.kn.term?.names?.[0])} */}
-      <span>{prefix}</span>
       <div id={adj.kn.uid}>
-        <div className="inline-flex items-start gap-2">
-          {adj.kn.term?.names?.map((name) => (
-            <span
-              key={name}
-              className="rounded-full font-bold text-green-800  dark:text-green-500"
-            >
-              {name}
-            </span>
-          ))}
-        </div>
-        {adj.kn.term?.names?.length && ":  "}
-        {toFormulas(adj.kn.sentence).map((formulaOrString) => {
-          if (typeof formulaOrString === "string") {
-            return (
-              <RefLinkSentence
-                key={formulaOrString}
-                sentence={formulaOrString}
-                refers={refs}
-              />
-            );
-          }
-
-          return formulaOrString;
-        })}
-        {adj.kn.additional && Object.keys(adj.kn.additional).length && (
-          <span className="inline-flex ml-2 text-sm text-muted-foreground">
-            <AdditionalComponent additional={adj.kn.additional} />
-          </span>
-        )}
-        <Relations startId={adj.kn.uid} />
+        <DefPresenter adj={adj} prefix={prefix} />
       </div>
     </div>
-  );
-}
-
-function forDisplay(value: string | object) {
-  if (typeof value === "string") {
-    return value;
-  }
-  const { terms, uids } = useResourceDetail();
-  if (value && "uid" in value && typeof value.uid === "string") {
-    const term = terms[value.uid];
-    const sentence = uids[value.uid];
-    const s =
-      // @ts-ignore
-      sentence && "n" in sentence
-        ? sentence?.n === "<<<not defined>>>"
-          ? ""
-          : sentence?.n
-        : sentence;
-    return `${term?.names?.join(", ")}: ${s}`;
-  }
-
-  throw new Error("invalid value");
-}
-
-function AdditionalComponent({ additional }: { additional: Additional }) {
-  if (!additional) {
-    return null;
-  }
-
-  const { when, where, by } = additional || {};
-  return (
-    <>
-      {when && (
-        <span className="flex items-center">
-          <Calendar className="size-4" />
-          {forDisplay(when)}
-        </span>
-      )}
-      {where && (
-        <span className="flex items-center">
-          <MapPin className="size-4" />
-          {forDisplay(where)}
-        </span>
-      )}
-      {by && (
-        <span className="flex items-center">
-          <User className="size-4" />
-          {forDisplay(by)}
-        </span>
-      )}
-    </>
   );
 }
